Finished adding global error handler to backend

+++ server/routes/projects.js
@@ -14,6 +14,7 @@ const {
 const createProjectSchema = require('../middleware/validation/projects/createProjectSchema.js');
 const updateProjectSchema = require('../middleware/validation/projects/updateProjectSchema.js');
 const deleteProjectSchema = require('../middleware/validation/projects/deleteProjectSchema.js');
+const { CustomError } = require('../utils/classes.js');
 
 const router = Router();
 
@@ -24,23 +25,22 @@ router.post(
 	'/create',
 	authenticateToken,
 	[checkSchema(createProjectSchema), schemaErrorHandler],
-	async (req, res) => {
-		const data = matchedData(req);
-		const { name, description } = data;
-
-		// Declared in authenticateToken middleware
-		const account_id = res.locals.account_id;
-
-		if (account_id == null) {
-			console.log('res.locals missing account_id');
-			return res.status(500).json({ errors: { server: 'Server error' } });
-		}
-
+	async (req, res, next) => {
 		try {
+			const data = matchedData(req);
+			const { name, description } = data;
+
+			// Declared in authenticateToken middleware
+			const account_id = res.locals.account_id;
+
+			if (account_id == null) {
+				throw new Error('res.locals missing account_id');
+			}
+
 			const createdProject = await pool.query(
 				`INSERT INTO project (account_id, name, description)
-						  VALUES ($1, $2, $3)
-           RETURNING project_id`,
+								VALUES ($1, $2, $3)
+						 RETURNING project_id`,
 				[account_id, name, description]
 			);
 
@@ -51,14 +51,13 @@ router.post(
 			const projects = await getProjectsFromDB(account_id);
 
 			if (projects == null) {
-				console.log('getProjectsFromDB returned without projects');
-				return res.status(500).json({ errors: { server: 'Server error' } });
+				throw new Error('getProjectsFromDB returned without projects array');
 			}
 
 			return res.status(200).json({ projects: projects.rows });
 		} catch (err) {
-			console.log(err.message);
-			return res.status(503).json({ errors: { server: 'Server error' } });
+			err.message = `create-project: ${err.message}`;
+			next(err);
 		}
 	}
 );
@@ -70,39 +69,34 @@ router.post(
 	'/update',
 	authenticateToken,
 	[checkSchema(updateProjectSchema), schemaErrorHandler],
-	async (req, res) => {
-		const data = matchedData(req);
-		const { project_id, name, description } = data;
-
-		// Declared in authenticateToken middleware
-		const account_id = res.locals.account_id;
-
-		if (account_id == null) {
-			console.log('res.locals missing account_id');
-			return res.status(500).json({ errors: { server: 'Server error' } });
-		}
-
+	async (req, res, next) => {
 		try {
-			const belongs = await doesProjectBelongToAccountInDB(
+			const data = matchedData(req);
+			const { project_id, name, description } = data;
+
+			// Declared in authenticateToken middleware
+			const account_id = res.locals.account_id;
+
+			if (account_id == null) {
+				throw new Error('res.locals missing account_id');
+			}
+
+			const projectBelongs = await doesProjectBelongToAccountInDB(
 				account_id,
 				project_id
 			);
-			if (!belongs) {
-				return res.status(400).json({
+
+			if (!projectBelongs) {
+				throw new CustomError('Project ID does not belong to account', 403, {
 					errors: { project_id: 'Project ID does not belong to account' },
 				});
 			}
-		} catch (err) {
-			console.log(err.message);
-			return res.status(503).json({ errors: { server: 'Server error' } });
-		}
 
-		try {
 			const updatedProject = await pool.query(
 				`UPDATE project
-				    SET name = $1, description = $2
-				  WHERE account_id = $3 AND project_id = $4
-				 RETURNING project_id`,
+							SET name = $1, description = $2
+						WHERE account_id = $3 AND project_id = $4
+					 RETURNING project_id`,
 				[name, description, account_id, project_id]
 			);
 
@@ -113,14 +107,13 @@ router.post(
 			const projects = await getProjectsFromDB(account_id);
 
 			if (projects == null) {
-				console.log('getProjectsFromDB returned without projects');
-				return res.status(500).json({ errors: { server: 'Server error' } });
+				throw new Error('getProjectsFromDB returned without projects');
 			}
 
 			return res.status(200).json({ projects: projects.rows });
 		} catch (err) {
-			console.log(err.message);
-			return res.status(503).json({ errors: { server: 'Server error' } });
+			err.message = `update-project: ${err.message}`;
+			next(err);
 		}
 	}
 );
@@ -132,37 +125,32 @@ router.delete(
 	'/delete',
 	authenticateToken,
 	[checkSchema(deleteProjectSchema), schemaErrorHandler],
-	async (req, res) => {
-		const data = matchedData(req);
-		const { project_id } = data;
-
-		// Declared in authenticateToken middleware
-		const account_id = res.locals.account_id;
-
-		if (account_id == null) {
-			console.log('res.locals missing account_id');
-			return res.status(500).json({ errors: { server: 'Server error' } });
-		}
-
+	async (req, res, next) => {
 		try {
-			const belongs = await doesProjectBelongToAccountInDB(
+			const data = matchedData(req);
+			const { project_id } = data;
+
+			// Declared in authenticateToken middleware
+			const account_id = res.locals.account_id;
+
+			if (account_id == null) {
+				throw new Error('res.locals missing account_id');
+			}
+
+			const projectBelongs = await doesProjectBelongToAccountInDB(
 				account_id,
 				project_id
 			);
-			if (!belongs) {
-				return res.status(400).json({
+
+			if (!projectBelongs) {
+				throw new CustomError('Project ID does not belong to account', 403, {
 					errors: { project_id: 'Project ID does not belong to account' },
 				});
 			}
-		} catch (err) {
-			console.log(err.message);
-			return res.status(503).json({ errors: { server: 'Server error' } });
-		}
 
-		try {
 			const deletedProject = await pool.query(
 				`DELETE FROM project
-				  WHERE account_id = $1 AND project_id = $2`,
+						WHERE account_id = $1 AND project_id = $2`,
 				[account_id, project_id]
 			);
 
@@ -173,14 +161,13 @@ router.delete(
 			const projects = await getProjectsFromDB(account_id);
 
 			if (projects == null) {
-				console.log('getProjectsFromDB returned without projects');
-				return res.status(500).json({ errors: { server: 'Server error' } });
+				throw new Error('getProjectsFromDB returned without projects');
 			}
 
 			return res.status(200).json({ projects: projects.rows });
 		} catch (err) {
-			console.log(err.message);
-			return res.status(503).json({ errors: { server: 'Server error' } });
+			err.message = `delete-project: ${err.message}`;
+			next(err);
 		}
 	}
 );
